pyproject.toml
[tool.poetry]
name = "report-generator"
version = "0.1.0"
description = ""
authors = ["vlad <vladislav.grachev@gmail.com>"]
readme = "README.md"
packages = [{include = "report_generator"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^20.7"}
python-dotenv = "^1.0.0"
google-api-python-client = "^2.111.0"
borb = "^2.1.20"
pydantic = {extras = ["email"], version = "^1.10.8"}
pandas = "^2.1.4"
redis = "^5.0.1"
bs4 = "^0.0.2"
lxml = "^5.1.0"
django = "^5.0.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
pre-commit = "^3.6.0"
poetry-plugin-export = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
